requirements.txt
Flask
python-dotenv 